feat(src): add no compositor mode functionality
Whenever you run the application in a WM that does not use a compositor,
the WA_TranslucentBackground from PyQt does not function properly and
simply renders a black fullscreen rectangle instead of a tinted
transparent one.

To better accomodate for those that don't use a compositor, such as
myself, if you have the option set to True, this logic will resolve the
black rectangle issue.

I would like to revisit this at some point and see if there is a better
solution, but for now this will work.