refactor(src:screendumper_overlay.py): optimize cleanup
This is a two for one combo, it solves the issue of temporary files
sticking around instead of being removed, when using the save to folder
function as well as making the cleanup and exit logic more versatile. I
felt it made more sense to bake the removal part into the cleanup logic
instead of duplicating that part of the code twice.